feat(Swagger): create Swagger and improve code

+++ src/controllers/orders.controller.js
@@ -2,11 +2,12 @@ import { Router } from 'express';
 import ordersService from '../services/orders.service.js';
 import productsService from '../services/products.service.js';
 import authMiddleware from '../middlewares/auth.middleware.js';
+import { transporter } from '../config/emails.service.js';
 import { sendCreatedOrder, sendUpdatedOrderStatus } from '../config/emails.service.js';
 import { isCustomerOrder } from '../middlewares/isCustomerOrder.middleware.js';
 import redisService from '../config/redis.service.js';
 import { validateInput } from '../middlewares/validateInput.middleware.js';
-import { updaetOrderSchema, updateOrderStatusSchema } from '../utils/enter.validators.js';
+import { updaetOrderSchema, updateOrderStatusSchema } from '../utils/inputData.validators.js';
 
 class OrdersController {
   path = '/order';
@@ -18,14 +19,14 @@ class OrdersController {
 
   setRotes() {
     this.router.route(`${this.path}/:id`).get(authMiddleware, isCustomerOrder, this.findOrderById);
-    this.router.route(`${this.path}`).get(authMiddleware, isCustomerOrder, this.findAllOrders);
+    this.router.route(`${this.path}`).get(authMiddleware, this.findAllOrders);
     this.router.route(`${this.path}/new`).post(authMiddleware, this.createOrder);
     this.router
       .route(`${this.path}/:id/update`)
       .patch(authMiddleware, isCustomerOrder, validateInput(updaetOrderSchema), this.updateOrder);
     this.router.route(`${this.path}/:id/delete`).delete(authMiddleware, isCustomerOrder, this.deleteOrder);
     this.router
-      .route(`${this.path}/status/:id/update`)
+      .route(`${this.path}/:id/update/status`)
       .patch(authMiddleware, isCustomerOrder, validateInput(updateOrderStatusSchema), this.updateOrderStatus);
   }
 
@@ -68,24 +69,26 @@ class OrdersController {
       const cachedCart = await redisService.getValue('cart');
       let cart = req.cookies.cart || JSON.parse(cachedCart) || [];
       let subTotalPrice = 0;
-      for (const item of cart) {
-        const product = await productsService.findProductById(item.productId);
-        subTotalPrice += product.price * item.quantity;
-      }
-      const newOrder = await ordersService.createOrder({
-        customer: req.user,
-        products: cart.map((item) => ({
-          product: item.productId,
-          quantity: item.quantity
-        })),
-        subTotalPrice
-      });
-      if (newOrder) {
-        await redisService.setValue(`order:${newOrder.id}`, JSON.stringify(newOrder));
-        await redisService.deleteValue('cart');
-        res.clearCookie('cart');
-        await sendCreatedOrder(newOrder);
-        return res.status(201).json(newOrder);
+      if (cart) {
+        for (const item of cart) {
+          const product = await productsService.findProductById(item.productId);
+          subTotalPrice += product.price * item.quantity;
+        }
+        const newOrder = await ordersService.createOrder({
+          customer: req.user,
+          products: cart.map((item) => ({
+            product: item.productId,
+            quantity: item.quantity
+          })),
+          subTotalPrice
+        });
+        if (newOrder) {
+          await redisService.setValue(`order:${newOrder.id}`, JSON.stringify(newOrder));
+          await redisService.deleteValue('cart');
+          res.clearCookie('cart');
+          transporter.sendMail(sendCreatedOrder(newOrder));
+          return res.status(201).json(newOrder);
+        }
       }
     } catch (error) {
       return res.status(error.status || 500).json({ error: error.message });
@@ -101,10 +104,10 @@ class OrdersController {
         const product = await productsService.findProductById(item.product.id);
         subTotalPrice += product.price * item.quantity;
       }
-      await ordersService.updateOrderSubTotalPrice(id, subTotalPrice);
-      if (updatedOrder) {
-        await redisService.setValue(`order:${id}`, JSON.stringify(updatedOrder));
-        return res.status(200).json(updatedOrder);
+      const updateOrder = await ordersService.updateOrderSubTotalPrice(id, subTotalPrice);
+      if (updateOrder) {
+        await redisService.setValue(`order:${id}`, JSON.stringify(updateOrder));
+        return res.status(200).json(updateOrder);
       }
       return res.status(404).json(`Order with id ${id} not found`);
     } catch (error) {
@@ -131,7 +134,7 @@ class OrdersController {
       const { id } = req.params;
       const updatedOrder = await ordersService.updateOrderStatus(id, req.body.status);
       if (updatedOrder) {
-        await sendUpdatedOrderStatus(updatedOrder);
+        await transporter.sendMail(await sendUpdatedOrderStatus(updatedOrder));
         await redisService.setValue(`order:${id}`, JSON.stringify(updatedOrder));
         return res.status(200).json(updatedOrder);
       }
